Refactor dos mixins de teste
